Fix bug with symbols logic in the webpage
When a button in webpage is pressed, the function writes the content of player object. But pressing the button another time fires the function again, thug changing the symbol. We don't want this behaviour to happen, so a conditional statement is made in the eventListener of the buttons. The buttons didn't fire the condition anyway because the buttons textContent needs to be changed with a spacew character to be recognised(otherwise it's flagged as undefined). A for loop before the execution of the eventListener solves the issue